None

+++ tests/test_prep_vasp.py
@@ -101,9 +101,10 @@ class TestPrepVaspDpConf(unittest.TestCase):
         for ii in [self.incar,self.potcar]:
             if Path(ii).is_file:
                 os.remove(ii)
-        for ii in self.confs:
-            if ii.is_dir():
-                shutil.rmtree(ii)
+        if not self.confs:
+            for ii in self.confs:
+                if ii.is_dir():
+                    shutil.rmtree(ii)
 
     def test(self):
         op = PrepVasp()
@@ -153,9 +154,10 @@ class TestPrepRunVaspPoscarConf(unittest.TestCase):
         for ii in [self.incar,self.potcar,self.optional_file]:
             if ii.is_file:
                 os.remove(ii)
-        for ii in self.confs:
-            if ii.is_file:
-                os.remove(ii)
+        if not self.confs:
+            for ii in self.confs:
+                if ii.is_file:
+                    os.remove(ii)
 
     def test(self):
         wf = Workflow(name = "test")
